+++ newlib/libc/stdlib/meson.build
@@ -33,20 +33,19 @@
 # OF THE POSSIBILITY OF SUCH DAMAGE.
 #
 nano_malloc_srcs_stdlib = [
-  'nano-malloc-calloc.c',
-  'nano-malloc-cfree.c',
-  'nano-malloc-free.c',
-  'nano-malloc-getpagesize.c',
-  'nano-malloc-mallinfo.c',
-  'nano-malloc-malloc.c',
-  'nano-malloc-malloc_stats.c',
-  'nano-malloc-malloc_usable_size.c',
-  'nano-malloc-mallopt.c',
-  'nano-malloc-memalign.c',
-  'nano-malloc-posix_memalign.c',
-  'nano-malloc-pvalloc.c',
-  'nano-malloc-realloc.c',
-  'nano-malloc-valloc.c',
+  'nano-calloc.c',
+  'nano-free.c',
+  'nano-getpagesize.c',
+  'nano-mallinfo.c',
+  'nano-malloc.c',
+  'nano-malloc-stats.c',
+  'nano-malloc-usable-size.c',
+  'nano-mallopt.c',
+  'nano-memalign.c',
+  'nano-posix-memalign.c',
+  'nano-pvalloc.c',
+  'nano-realloc.c',
+  'nano-valloc.c',
 ]
 
 std_malloc_srcs_stdlib = [
